Moved set_poes in contexts.py
